Dev/1.0dev2

+++ jsqp_core/packages/file_package.py
@@ -1,13 +1,17 @@
 from __future__ import annotations
+from typing import TYPE_CHECKING
 
 import os
 import shutil
-from io import FileIO
+import hashlib
 from enum import Enum
 from pathlib import Path
 from zipfile import ZipFile
-from abc import abstractmethod
+from abc import abstractmethod, ABC
 from devgoldyutils import LoggerAdapter
+
+if TYPE_CHECKING:
+    from io import BufferedReader
 
 from ..paths import Paths
 from .package import Package
@@ -23,7 +27,7 @@ class FileTypes(Enum):
     SYMBOLIC_LINK = 2
     ZIP = 3
 
-class FilePackage(Package):
+class FilePackage(Package, ABC):
     """A package represented as an actual file."""
     def __init__(
         self, 
@@ -42,7 +46,14 @@ class FilePackage(Package):
             self.logger.debug("Extracting zip to temp directory...")
             temp_path = Paths().temp_dir + f"/{path.name[:-4]}"
 
-            ZipFile(path.absolute()).extractall(temp_path)
+            try:
+                ZipFile(path.absolute()).extractall(temp_path)
+            except Exception as e:
+                raise errors.JSQPCoreError(
+                    "This error is most likely NOT a bug in jsqp! It seems ZipFile couldn't extract your package. " \
+                    "This could mean your package is corrupted in some way or another. " \
+                    f"Error >> {e}"
+                )
 
             path = Path(temp_path)
 
@@ -60,8 +71,8 @@ class FilePackage(Package):
 
     @property
     @abstractmethod
-    def install_location(self) -> str:
-        """Returns the jsqp install location for this file package."""
+    def package_install_path(self) -> str:
+        """Location of where this package is supposed to be installed. NOT the actual installed location."""
         return Paths().jsqp_core_appdata_dir + "/packages"
 
     @property
@@ -76,19 +87,16 @@ class FilePackage(Package):
         self.logger.debug(f"Updated file package path object to '{str(value)}'!")
 
     @property
-    def file(self) -> FileIO | None:
+    def file(self) -> BufferedReader | None:
         """Returns the file object of this package. Returns none if this package isn't a file."""
         if self.__path.is_file():
-            file = open(str(self.__path), "r")
-            file.close()
-            return file
+            return open(str(self.__path), "rb")
 
         return None
 
     @property
     def type(self) -> FileTypes:
         """Returns the type of FilePackage. Is it a folder? Is it a file?"""
-
         if self.path.is_file():
             if self.path.suffix == ".zip":
                 return FileTypes.ZIP
@@ -100,24 +108,38 @@ class FilePackage(Package):
         elif self.path.is_symlink():
             return FileTypes.SYMBOLIC_LINK
 
+    @property
+    def hash(self) -> str | None:
+        """Returns the hash of the package's file. Returns None if self.file is none."""
+        file = self.file
+
+        if file is not None:
+            md5 = hashlib.md5()
+
+            with file as f:
+                md5.update(f.read())
+
+            return md5.hexdigest()
+
+        return None
+
     def add(self, overwrite: bool = False):
-        """Adds the package to the jsqp repository."""
-        self.logger.info(f"Adding '{self.display_name}' to jsqp repository...")
+        """Adds the package to your jsqp storage."""
+        self.logger.info(f"Adding '{self.display_name}' to jsqp 📦 storage...")
+
         # Zip it if it's not already a zip.
         if self.type == FileTypes.FOLDER:
-            self.zip(self.name + ".zip", config.performance_mode)
+            self.zip(self.name + ".zip")
 
         # Move the package to JSQPCore install location directory.
-        try:
-            self.move(self.install_location, overwrite, config.no_copy)
-        except FileNotFoundError:
-            # Repair and try again if file is not found.
-            # -------------------------------------------
-            Paths().repair_app_data_dir([
-                self.install_location
-            ])
 
-            self.move(self.install_location, overwrite, config.no_copy)
+        if not os.path.exists(self.package_install_path):
+            Paths().repair_app_data_dir([self.package_install_path])
+
+        try:
+            self.move(
+                self.package_install_path, overwrite, config.no_copy
+            )
         except FileExistsError:
             raise errors.PackageAlreadyExist(self)
 
@@ -137,26 +159,29 @@ class FilePackage(Package):
             os.link(self.path.absolute(), target_location)
             self.logger.debug(f"Created symbolic link '{self.name}' at '{location}'.")
         except FileExistsError:
+            # TODO: We probably would want to overwrite it instead of causing an exception.
             self.logger.warning("Didn't create symbolic link as it already exists.")
 
     def move(self, path: str, overwrite: bool = False, no_copy: bool = True) -> None:
-        """Allows you to move this file package to another location. Raises ``FileNotFoundError`` if path or file does not exist."""
+        """
+        Allows you to move this file package to another location. 
+        Raises ``FileNotFoundError`` if path does not exist and ``FileExistsError`` if a file already exists there.
+        """
         new_path = f"{path}/{self.path.name}"
         self.logger.info(f"Moving '{self.display_name}' to '{os.path.split(new_path)[0]:.60}'...")
 
-        if overwrite:
-            self.logger.info("Overwriting as it already exists...")
-            if os.path.isdir(new_path):
-                shutil.rmtree(new_path, True)
-            else:
-                os.remove(new_path)
-
         if os.path.exists(new_path):
-            raise FileExistsError("The file already exists.")
+
+            if overwrite or config.overwrite:
+                self.logger.debug("Overwriting as it already exists...")
+                shutil.rmtree(new_path, True) if os.path.isdir(new_path) else os.remove(new_path)
+
+            else:
+                raise FileExistsError("The file already exists.")
 
         shutil.copy2(self.path.absolute(), new_path)
 
-        # Delete if the user only doesn't want to copy the file.
+        # Delete if the user only wants to copy the file.
         if no_copy is False:
             self.delete()
 
@@ -165,7 +190,7 @@ class FilePackage(Package):
 
         self.logger.info(f"Moved to '{new_path:.60}'!")
 
-    def zip(self, zip_name: str = None, performance_mode: bool = False) -> bool:
+    def zip(self, zip_name: str = None) -> bool:
         """Turn file package into a zip if it is a folder."""
         if zip_name is None:
             zip_name = self.name + ".zip"
@@ -175,22 +200,13 @@ class FilePackage(Package):
             path_to_zip = Path(str(self.path.parent) + f"/{zip_name}")
 
             self.logger.info(f"Zipping {self.display_name} to '{str(path_to_zip):.60}...'")
-            
-            # Zipping each file in directory.
-            # --------------------------------
-            with ZipFile(path_to_zip.absolute(), mode = "w") as archive:
 
-                if core_logger.level == 10 and performance_mode is False: # DEBUG MODE!
-                    for file_path in self.path.rglob("*"):
-                        archive.write(file_path, arcname = file_path.__str__().partition(self.path.name + os.path.sep)[2])
-                        self.logger.debug(f"Zipped '{file_path.name}'.")
-                
-                else: # If debug logging level is not set we can zip much FASTER!
-                    for file_path in self.path.rglob("*"):
-                        archive.write(file_path, arcname = file_path.__str__().partition(self.path.name + os.path.sep)[2])
+            with ZipFile(path_to_zip.absolute(), mode = "w") as archive:
+                for file_path in self.path.rglob("*"):
+                    archive.write(file_path, arcname = file_path.__str__().partition(self.path.name + os.path.sep)[2])
 
             self.logger.debug(f"Done zipping to '{path_to_zip.absolute()}'!")
-            
+
             # Updating path object.
             self.path = path_to_zip
